+++ src/ardrive.test.ts
@@ -15,7 +15,7 @@ import { ARDataPriceRegressionEstimator } from './pricing/ar_data_price_regressi
 import { GatewayOracle } from './pricing/gateway_oracle';
 import { ByteCount, UnixTime, stubTransactionID, W, FeeMultiple, TipType } from './types';
 import { readJWKFile } from './utils/common';
-import { stubEntityID } from './utils/stubs';
+import { stubEntityID } from '../tests/stubs';
 import { expectAsyncErrorThrow } from '../tests/test_helpers';
 import { WalletDAO } from './wallet_dao';
