+++ src/ardrive.ts
@@ -73,10 +73,10 @@ import {
 } from './types';
 import { urlEncodeHashKey } from './utils/common';
 import { errorMessage } from './utils/error_message';
-import { stubEntityID } from './utils/stubs';
 import { Wallet } from './wallet';
 import { JWKWallet } from './jwk_wallet';
 import { WalletDAO } from './wallet_dao';
+import { fakeEntityId } from './utils/constants';
 
 export class ArDrive extends ArDriveAnonymous {
 	constructor(
@@ -1097,7 +1097,7 @@ export class ArDrive extends ArDriveAnonymous {
 		// Assert that there's enough AR available in the wallet
 		// Use stub data to estimate costs since actual data requires entity IDs generated by ArFSDao
 		const stubRootFolderData = new ArFSPublicFolderTransactionData(driveName);
-		const stubDriveData = new ArFSPublicDriveTransactionData(driveName, stubEntityID);
+		const stubDriveData = new ArFSPublicDriveTransactionData(driveName, fakeEntityId);
 		const driveUploadCosts = await this.estimateAndAssertCostOfDriveCreation(stubDriveData, stubRootFolderData);
 		const driveRewardSettings = {
 			reward: driveUploadCosts.driveMetaDataBaseReward,
@@ -1140,7 +1140,7 @@ export class ArDrive extends ArDriveAnonymous {
 		const stubRootFolderData = await ArFSPrivateFolderTransactionData.from(driveName, newPrivateDriveData.driveKey);
 		const stubDriveData = await ArFSPrivateDriveTransactionData.from(
 			driveName,
-			stubEntityID,
+			fakeEntityId,
 			newPrivateDriveData.driveKey
 		);
 		const driveCreationCosts = await this.estimateAndAssertCostOfDriveCreation(stubDriveData, stubRootFolderData);
@@ -1514,10 +1514,10 @@ export class ArDrive extends ArDriveAnonymous {
 			lastModifiedDateMS,
 			stubTransactionID,
 			dataContentType,
-			stubEntityID,
+			fakeEntityId,
 			await deriveDriveKey(
 				'stubPassword',
-				`${stubEntityID}`,
+				`${fakeEntityId}`,
 				JSON.stringify((this.wallet as JWKWallet).getPrivateKey())
 			)
 		);
