+++ src/exports.ts
@@ -53,5 +53,4 @@ export * from './utils/mapper_functions';
 export * from './utils/query';
 export * from './utils/crypto';
 export * from './utils/sort_functions';
-export * from './utils/stubs';
 export * from './utils/wallet_utils';
