+++ src/utils/constants.ts
@@ -1,5 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { version: DEFAULT_APP_VERSION } = require('../../package.json');
+import { EID } from '../types';
 import { CipherType } from '../types/type_guards';
 
 export const ArFS_O_11 = '0.11';
@@ -18,6 +19,8 @@ export const appName = 'ArDrive-Desktop';
 export const webAppName = 'ArDrive-Web';
 
 export const defaultCipher: CipherType = 'AES256-GCM';
+
+export const fakeEntityId = EID('00000000-0000-0000-0000-000000000000');
 
 /**
  * Minimum ArDrive community tip from the Community Improvement Proposal Doc:
